// PaginaFinal/app/src/types.ts
export type PageId =
  | 'dashboard'
  | 'mi-espacio'
  | 'perfil'
  | 'grupos'
  | 'alumnos'
  | 'docentes'
  | 'carreras'
  | 'materias'
  | 'calificaciones'
  | 'asistencia'
  | 'horarios'
  | 'inscripciones'
  | 'kardex'
  | 'biblioteca'
  | 'inventarios'
  | 'plataforma-trabajos'
  | 'servicios'
  | 'reportes'
  | 'configuracion'

export type Role = 'Administrador' | 'Control Escolar' | 'Docente' | 'Alumno' | 'Vinculacion'

export type ViewMode = 'public' | 'login' | 'app'